requirements.txt
python-telegram-bot==21.6
uvicorn==0.30.6
fastapi==0.114.2
pydantic==2.9.2
httpx==0.27.2
pyyaml==6.0.2
